requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
decorator==5.1.1
Django==4.0.2
djangorestframework==3.13.1
executing==0.8.2
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
ipython==8.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
numpy==1.22.2
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
ratelim==0.1.6
requests==2.27.1
six==1.16.0
sodapy==2.1.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
